added entriesController.scala
added entriesController.scala (wip) and fixed entries.scala

+++ app/models/entries.scala
@@ -1,6 +1,6 @@
 package models
 import java.time
 
-case class entries(id: Int, userId: Int, date: LocalDateTime, title: String, body: String)
+case class entries(id: Long, userId: Long, date: LocalDateTime, title: String, body: String)
 
 case class entriesDto(date: LocalDateTime, title: String, body: String)
